Phase 2 audit remediation: break the coordinator cycle, decouple architecture
2.1 Manager->Coordinator cycle broken via NotchUIEventBus:
- Managers (Volume/Brightness/KeyboardBacklight/BetterDisplay/Lunar/
  Music/Battery) now publish presentation events; the coordinator is the
  single subscriber and keeps all show/hide policy. Managers are now
  testable without the UI stack.
- Removed nested @ObservedObject coordinator from MusicManager and
  BatteryStatusViewModel (kills whole-tree republishing of both).

2.2 Defaults<->MusicManager init cycle removed:
- New MediaEnvironment owns the NowPlaying availability probe (resolved
  eagerly at launch, persisted for static key defaults).
- Defaults.Keys.mediaController default reads only UserDefaults — can no
  longer re-enter MusicManager's lazy init (latent fatal error).

2.5 XPC contract single-sourced + helper health:
- Both targets now compile the SAME files via a Shared/ synchronized
  group: BoringNotchXPCHelperProtocol.swift + unified JSONLinesPipeHandler
  (the app/helper copies had drifted; helper also lost its dead NSCoder init).
- New XPCHelperError type; client tracks connection health and Settings
  shows a 'Helper Service Unavailable' warning when the helper dies
  instead of features silently degrading.

2.3 AppDelegate god object split (671 -> ~330 lines):
- New NotchWindowManager owns window/view-model/drag-detector lifecycle;
  parallel [UUID] dictionaries unified into one ScreenContext per screen.
- Fixes latent leak: windowScreenDidChange observers are now removed
  before re-registration (used to leak one per window recreation).

2.4 ShelfItemViewModel god object split (1155 -> 182 lines):
- AppKit context-menu construction + dispatch extracted to
  ShelfContextMenu.swift via closure bridge (no singleton back-ref).
- Menu dispatch is now tag-based (ContextMenuAction rawValue) instead of
  matching NSMenuItem.title — which silently broke under any non-English
  locale, and would have broken as soon as titles were localized.
- Menu strings now use the file's existing translated Strings table
  (Shelf.ContextMenu.* keys present in the catalog) instead of literals.
- Removed dead loadOpenWithApps.

2.6 Concurrency hygiene:
- MusicManager is now @MainActor (was publishing from arbitrary tasks).
- WebcamManager: removed @Published-duplicate didSet objectWillChange sends.

Verified: Debug build succeeds for app + XPC helper.
