+++ boringNotch/MediaControllers/NowPlayingStreamSupport.swift
@@ -147,62 +147,6 @@ enum NowPlayingAvailability: Equatable, Sendable {
     }
 }
 
-actor JSONLinesPipeHandler {
-    nonisolated let outputPipe: Pipe
-    nonisolated let fileHandle: FileHandle
-    private var byteIterator: FileHandle.AsyncBytes.Iterator
-    private var consecutiveMalformedLines = 0
-
-    init(pipe: Pipe = Pipe()) {
-        outputPipe = pipe
-        fileHandle = pipe.fileHandleForReading
-        byteIterator = fileHandle.bytes.makeAsyncIterator()
-    }
-
-    func readJSONLines<Value: Decodable & Sendable>(
-        as type: Value.Type,
-        onValue: @escaping @Sendable (Value) async -> Void
-    ) async {
-        var line = Data()
-        var iterator = byteIterator
-
-        do {
-            while let byte = try await iterator.next() {
-                guard !Task.isCancelled else { return }
-
-                guard byte == UInt8(ascii: "\n") else {
-                    line.append(byte)
-                    continue
-                }
-
-                if line.last == UInt8(ascii: "\r") {
-                    line.removeLast()
-                }
-
-                guard !line.isEmpty,
-                      let decoded = try? JSONDecoder().decode(Value.self, from: line)
-                else {
-                    consecutiveMalformedLines += 1
-                    line.removeAll(keepingCapacity: true)
-                    if consecutiveMalformedLines >= 3 {
-                        return
-                    }
-                    continue
-                }
-
-                consecutiveMalformedLines = 0
-                line.removeAll(keepingCapacity: true)
-                await onValue(decoded)
-            }
-        } catch {}
-    }
-
-    nonisolated func close() {
-        try? fileHandle.close()
-        try? outputPipe.fileHandleForWriting.close()
-    }
-}
-
 @MainActor
 final class NowPlayingStreamSession {
     private let process: Process
